requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30       
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1      
Flask-Cors==5.0.0        
Flask-Login==0.6.3       
Flask-PyMongo==2.3.0
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
maclookup==1.0.3
maclookup-cli==0.1.9
MarkupSafe==2.1.5
netaddr==1.3.0
ply==3.11
psutil==6.0.0
pyasn1==0.6.1
pymongo==4.9.1
pysmi==1.5.0
pysnmp==7.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nmap==0.7.1
requests==2.32.3
six==1.16.0
speedtest-cli==2.1.3
urllib3==2.2.3
Werkzeug==3.0.4


# Use the following command to install the dependencies:
# pip install -r requirements.txt
